Clean up Tracking::PreintegrateIMU() preamble
This function begins by retrieving relevant samples from mlQueueImuData.
This commit makes three changes to that preamble.

It looks like a previous version used to sleep for 500us at a time if no
samples were available in the queue, but currently it's impossible for
bSleep to become true. Remove the dead code.

Perform all access to mlQueueImuData under the protection of
mMutexImuQueue, as even const member functions like std::list::size are
not guaranteed to be atomic.

Change the type of mlQueueImuData from std::list to std::queue for lower
overhead.